refactor(spectral tools): Reduce Au3 theme-related classes to strict minimum for Au4 spectrogram purpose
