+++ au3/libraries/au3-theme/Theme.h
@@ -14,56 +14,15 @@
 #ifndef __AUDACITY_THEME__
 #define __AUDACITY_THEME__
 
-#include <map>
-#include <unordered_set>
-#include <vector>
-#include <optional>
+#include "ComponentInterfaceSymbol.h" // Identifier
 
 #include <QColor>
 
-#include "au3-components/ComponentInterfaceSymbol.h"
+#include <wx/arrstr.h>
 
-#include "au3-utility/Observer.h"
-#include "au3-preferences/Prefs.h"
-
-//! A choice of theme such as "Light", "Dark", ...
-using teThemeType = Identifier;
-
-//! A system theme, that matches selected theme best (only works on macOS with builtin themes).
-enum class PreferredSystemAppearance
-{
-    Light,
-    Dark,
-    HighContrastDark
-};
-
-class wxArrayString;
-class wxBitmap;
-class wxColour;
-class wxImage;
-class wxPen;
-
-class ChoiceSetting;
-
-// JKC: will probably change name from 'teBmps' to 'tIndexBmp';
-using teBmps = int; /// The index of a bitmap resource in Theme Resources.
-
-enum teResourceType
-{
-    resTypeColour,
-    resTypeBitmap,
-    resTypeImage = resTypeBitmap,
-};
-
-enum teResourceFlags
-{
-    resFlagNone   =0x00,
-    resFlagPaired =0x01,
-    resFlagCursor =0x02,
-    resFlagNewLine = 0x04,
-    resFlagInternal = 0x08, // For image manipulation.  Don't save or load.
-    resFlagSkip = 0x10
-};
+#include <map>
+#include <unordered_set>
+#include <vector>
 
 struct ThemeSet
 {
@@ -73,13 +32,7 @@ struct ThemeSet
     bool bInitialised = false;
 };
 
-struct ThemeChangeMessage {
-    std::optional<PreferredSystemAppearance> appearance; /*!<
-      An enum value when preferred system appearance changes, or nullopt
-      for change of the image set */
-};
-
-class THEME_API ThemeBase /* not final */ : public Observer::Publisher<ThemeChangeMessage>
+class THEME_API ThemeBase
 {
 public:
     ThemeBase (const ThemeBase&) = delete;
@@ -115,9 +68,5 @@ public:
 };
 
 extern THEME_API Theme theTheme;
-
-extern THEME_API ChoiceSetting
-& GUITheme()
-;
 
 #endif // __AUDACITY_THEME__
